requirements.txt
llama-index-llms-openai
llama-index-readers-file
llama-index
pymilvus
llama-index-vector-stores-milvus
llama-index-vector-stores-chroma
milvus
h3pandas
chromadb
streamlit
openai
langchain
langchain-experimental
langchain-community
geopandas
folium
streamlit_folium
branca
tabulate
plotly_express